Api cleanup and refactoring
Removed unessassary routes and changed every frontend fetch operations to use an environment variable for the api url
